Move Guild classes to their own package
