+++ src/main/kotlin/Character.kt
@@ -5,11 +5,11 @@ import cloud.drakon.ktlodestone.exception.LodestoneException
 import cloud.drakon.ktlodestone.profile.ActiveClassJob
 import cloud.drakon.ktlodestone.profile.GrandCompany
 import cloud.drakon.ktlodestone.profile.Guardian
-import cloud.drakon.ktlodestone.profile.Guild
-import cloud.drakon.ktlodestone.profile.GuildType
-import cloud.drakon.ktlodestone.profile.IconLayers
 import cloud.drakon.ktlodestone.profile.ProfileCharacter
 import cloud.drakon.ktlodestone.profile.Town
+import cloud.drakon.ktlodestone.profile.guild.Guild
+import cloud.drakon.ktlodestone.profile.guild.GuildType
+import cloud.drakon.ktlodestone.profile.guild.IconLayers
 import io.ktor.client.call.body
 import io.ktor.client.request.get
 import kotlinx.coroutines.async
