+++ src/main/kotlin/profile/guild/GuildType.kt
@@ -1,4 +1,4 @@
-package cloud.drakon.ktlodestone.profile
+package cloud.drakon.ktlodestone.profile.guild
 
 internal enum class GuildType {
     FREE_COMPANY, PVP_TEAM
